test(desktop): move test-only helpers out of the shipped binary
Problem
Five production files in the desktop module carry helpers that only the test
suite calls: setSessionTitle and deleteSessionFile (sessions.go),
HeartbeatEngine.loadTasks and .saveTasks (heartbeat_store.go), historyMessages,
skillRootsView, mcpFailed and saveExclusiveExportFiles (app.go),
pathWithinWorktree (delivery_worktree.go) and buildScrollDiagnosticsZip
(scroll_diagnostics_export.go). They compile into the app binary and read as
production API.

Fix
Move all ten into desktop/testhelpers_test.go. No call site changes: the tests
keep the same names, and no production path referenced any of them. Moving
them into an existing test file was not an option - sessions_test.go and
app_test.go already sit above repolint's 800-line ceiling, so growing them
would create a new violation - hence one small file.

Verification
- go build ./... and go vet ./... clean in the desktop module
- go test ./... in the desktop module
- go run ./tools/repolint reports nothing for the touched files or the new one

Not moved
The icon_repair_darwin helpers (writeMacAlias, resolveMacAlias,
macURLReferencesReasonixBundle) stay where they are: they call cgo, and Go does
not allow import "C" in a _test.go file.
