+++ desktop/sessions.go
@@ -109,41 +109,6 @@ func saveSessionTitles(dir string, m map[string]string) error {
 	return fileutil.AtomicWriteFile(sessionTitlesPath(dir), b, 0o600)
 }
 
-// setSessionTitle sets (or, with an empty title, clears) a session's custom name.
-func setSessionTitle(dir, sessionPath, title string) error {
-	sessionPath, _, err := validateSessionPath(dir, sessionPath)
-	if err != nil {
-		return err
-	}
-	key := filepath.Base(sessionPath)
-	return updateSessionTitles(dir, func(m map[string]string) bool {
-		title = strings.TrimSpace(title)
-		if title == "" {
-			if _, ok := m[key]; !ok {
-				return false
-			}
-			delete(m, key)
-			return true
-		}
-		if m[key] == title {
-			return false
-		}
-		m[key] = title
-		return true
-	})
-}
-
-// deleteSessionFile moves a session's .jsonl and file sidecars into the local
-// trash. Title/display sidecars stay in place so trash previews and restores can
-// preserve the user's labels.
-func deleteSessionFile(dir, sessionPath string) error {
-	sessionPath, key, err := validateSessionPath(dir, sessionPath)
-	if err != nil {
-		return err
-	}
-	return trashSessionArtifacts(dir, sessionPath, key)
-}
-
 type trashedSessionMeta struct {
 	Key       string `json:"key"`
 	DeletedAt int64  `json:"deletedAt"`
